+++ transformers_lightning/optimizers/adamw_electra.py
@@ -103,8 +103,8 @@ class ElectraAdamWOptimizer(SuperOptimizer, ElectraAdamW):
             amsgrad=hyperparameters.amsgrad
         )
 
-    def add_optimizer_specific_args(parser: ArgumentParser):
-        super(ElectraAdamWOptimizer, ElectraAdamWOptimizer).add_optimizer_specific_args(parser)
+    def add_argparse_args(parser: ArgumentParser):
+        super(ElectraAdamWOptimizer, ElectraAdamWOptimizer).add_argparse_args(parser)
         parser.add_argument('--learning_rate', type=float, default=1e-3)
         parser.add_argument('--weight_decay', type=float, default=0.01)
         parser.add_argument('--adam_epsilon', type=float, default=1e-6)