+++ transformers_lightning/schedulers/constant_scheduler_with_warmup.py
@@ -50,7 +50,7 @@ class ConstantSchedulerWithWarmup(SuperScheduler):
         return [base_lr * self.lr_lambda(self.last_epoch) for base_lr in self.base_lrs]
 
     @staticmethod
-    def add_scheduler_specific_args(parser: ArgumentParser):
+    def add_argparse_args(parser: ArgumentParser):
         r""" Add here the hyperparameters specific of the scheduler like the number of warmup steps. """
-        super(ConstantSchedulerWithWarmup, ConstantSchedulerWithWarmup).add_scheduler_specific_args(parser)
+        super(ConstantSchedulerWithWarmup, ConstantSchedulerWithWarmup).add_argparse_args(parser)
         parser.add_argument('--num_warmup_steps', type=int, default=0)